refactor: Canvas Component from Whiteboard
