+++ client/app.tsx
@@ -2,7 +2,6 @@ import * as React from 'react'
 
 import Navbar from './components/navbar'
 import Routes from './routes'
-import { Whiteboard } from './components/whiteboard'
 
 export default class App extends React.Component {
   public render() {
