+++ client/components/navbar.js
@@ -15,14 +15,14 @@ const Navbar = ({handleClick, isLoggedIn}) => (
           <a href="#" onClick={handleClick}>
             Logout
           </a>
-          <Link to="/board">Board</Link>
+          <Link to="/canvas">Board</Link>
         </div>
       ) : (
         <div>
           {/* The navbar will show these links before you log in */}
           <Link to="/login">Login</Link>
           <Link to="/signup">Sign Up</Link>
-          <Link to="/board">Board</Link>
+          <Link to="/canvas">Board</Link>
         </div>
       )}
     </nav>
