+++ client/routes.js
@@ -4,7 +4,7 @@ import {withRouter, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
 import {Login, Signup, UserHome} from './components'
 import {me} from './store'
-import { Whiteboard } from './components/whiteboard'
+import { Canvas } from './components/canvas'
 
 /**
  * COMPONENT
@@ -22,12 +22,12 @@ class Routes extends Component {
         {/* Routes placed here are available to all visitors */}
         <Route path="/login" component={Login} />
         <Route path="/signup" component={Signup} />
-        <Route path="/board" component={Whiteboard} />
+        <Route path="/canvas" component={Canvas} />
         {isLoggedIn && (
           <Switch>
             {/* Routes placed here are only available after logging in */}
             <Route path="/home" component={UserHome} />
-            <Route path="/board" component={Whiteboard} />
+            <Route path="/canvas" component={Canvas} />
           </Switch>
         )}
         {/* Displays our Login component as a fallback */}
